Add log_remote_query GUC to control logging of remote MongoDB queries.
Introduce a new GUC log_remote_query to enable logging of the remote
query that would be executed on the MongoDB server.  This is intended
for debugging purposes only.  When enabled, the remote query is
printed to the PostgreSQL server logs.  The default value is false.

FDW-729, Vaibhav Dalvi, reviewed and some adjustments by Jeevan Chalke.